+++ src/checker/fileNamingPattern.ts
