+++ src/checker/strictFormatCheck.ts
